requirements.txt
openpyxl>=3.1.0
pandas>=2.0.0
pydantic>=2.0.0
python-pptx>=0.6.21
matplotlib>=3.7.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2
contextily>=1.4.0
geopandas>=0.14.0
anthropic>=0.39.0
Pillow>=10.0.0
click>=8.1.0
python-dotenv>=1.0.0
shapely>=2.0.0
flask>=3.0.0
